Remove vmsplice as it is an unsafe operation in python()
vmsplice does not copy data but maps data to the page, modifcations
to the original data are reflected in the pipe buffer up until the
read(). as we cant control the other end of the pipes() read and
when python does a gc this means that a gc cycle can occur before
a read() causing linux to read random data from other objects.

After much thorugh about this issue the interface/syscall we actually
want is writev(2) which is provided by the 'os' module in python.
its arguments are nearly 1:1 and it has the copying behavior we desire